requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
fastapi==0.103.2
h11==0.14.0
httptools==0.6.0
idna==3.4
pydantic==2.4.2
pydantic[email]
pydantic_core==2.10.1
pydantic-settings==2.0.3
python-dotenv==0.21.1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
watchfiles==0.20.0
websockets==11.0.3
sqlalchemy==2.0.21
passlib==1.7.4
fastapi-jwt==0.1.8
bcrypt==4.0.1
httpx==0.24.1
pytest==7.4.1
psycopg2-binary==2.9.9
python-multipart==0.0.5
easyocr==1.7.1
opencv-python-headless==4.8.1.78
scikit-learn==1.3.2
azure-storage-blob==12.9.0
